vote/requirements.txt
Flask
Redis
gunicorn
python-consul
